+++ src/rilog/__init__.py
@@ -2,4 +2,8 @@
 
 from importlib.metadata import version
 
+from rilog.logger import Logger, logger
+
 __version__ = version("rilog")
+
+__all__ = ["Logger", "logger"]