+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+## Unreleased
+
+- Introduced an enhanced logger with methods for logging messages and warnings, setting/removing a log prefix,
+  and tracking progress of data iteration with a progress bar.